Extract date operator helpers
